Fix dictionary key in YT-dl progress_hooks, Add condition before starting upload of downloaded files in YT-dl

+++ mega/helpers/ytdl.py
@@ -105,7 +105,10 @@ class YTdl:
                     file_name = os.path.basename(temp_file)
                     pre, _ = os.path.splitext(file_name)
                     temp_file = os.path.join(temp_dir, f"{pre}.mp3")
-                await UploadFiles().upload_file(temp_file, ack_message, msg.text, "other", "disk", "")
+                if os.path.isfile(temp_file):
+                    await UploadFiles().upload_file(temp_file, ack_message, msg.text, "other", "disk", "")
+                else:
+                    continue
                 break
             else:
                 await asyncio.sleep(1)
@@ -127,7 +130,7 @@ class YTdl:
                     int(d["downloaded_bytes"]), binary=True
                 )
                 yt_progress_updates[f"{chat_id}{message_id}"]["total"] = size.format_size(
-                    int(d["total_bytes"]), binary=True
+                    int(d["total_bytes_estimate"]), binary=True
                 )
                 yt_progress_updates[f"{chat_id}{message_id}"]["file_name"] = d["filename"]
         except Exception as e:
